Replace `DownloadDepot` with plain function
